Refactor CV and analysis and integrate into training loop
This change refactors the cross-validation and analysis functionality into dedicated classes, `CrossValidator` and `Analysis`, to improve modularity and create a more organized workflow.

The cross-validation is now integrated directly into the training loop in `train_model.py`. After each model is trained, it is immediately cross-validated in memory, which is more efficient as it avoids loading models from disk.

A new main script, `scripts/main.py`, orchestrates the entire process, from training and cross-validation to analysis, providing a single point of entry.

The original standalone scripts for cross-validation and analysis have been removed as they are now redundant.
